ai-service/requirements.txt
Flask==2.3.3
groq
flask-limiter
tenacity
bleach
python-dotenv
redis
pytest
flask-cors